feat(m25): complete decomposition extraction modules (#2032)
